[netstack3] Carry errors with serializers
Introduce ErrorWithSerializer and an alias to it, SendFrameError. This
encodes the pattern of including errors alongside serializers when
errors are observed on the send path.

Bubble all of it up through to the IP layer.

A follow-up CL will use the now better-placed errors to introduce
IpSendFrameError to scope errors to the IP layer, like invalid loopback
addresses on the wire.

Bug: 346585534
Fixed: 42167236

Change-Id: I34c313b60ec0ebb233a02d6c94d753e4b40aa397
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065958
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>