[tee] Improve handling of dead TA
When a TA instance dies the trusted OS is supposed to return
ERROR_TARGET_DEAD with a return origin indicating the trusted OS for
future calls on that application. This implements that logic in the TA
manager.

Bug: b/334930640
Change-Id: I91f56951b8e75b2b60c0f87d3d9621e37806573a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1066250
Reviewed-by: Eric Stone <ecstone@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>