[Netstack3] Use ResolvedRoute::local_delivery_device
resolve_route_to_destination() returns local_delivery_device as part
of ResolvedRoute. That field is intended to specify a device for
loopback delivery, but it was ignored during packet delivery. Now
it's passed to the loopback device as expected.
This required some fixes in resolve_route_to_destination - it was
using SRC IP to resolve the delivery device, which is incorrect.
Also renamed parameters in resolve_route_to_destination to make it
less error-prone.

Bug: 42084713
Change-Id: I6e634fd179bbb6a7dfc612b25694ddd952f2918d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063879
Commit-Queue: Sergey Ulanov <sergeyu@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>