[fx] Fix list-packages script
The list-packages tool can be used to output a list of packages
available in the current Fuchsia build. The list is currently sourced
from several build artifacts that are not guaranteed to be available
depending on how the build was assembled, and should instead be sourced
from all_package_manifests.list which is always present.

Changes:
- Package info for list-packages is now sourced from
  all_package_manifests.list
  - Note that this eliminates the ability to list packages based on
    whether they were base, universe, or cache packages; that ability
    relied on the filename of the build artifact the list was sourced
    from and is not contained within the actual manifest files for each
    package (which are now used to source package names)

Bug: b/337957623

Change-Id: Ic14d8205fc5c22c1be7c98b90e4087d76017be8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052925
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Alex Wong <alexdwong@google.com>