[kernel] Convert platform_set_oneshot_timer to take a raw ticks value
Modifies the platform_set_oneshot_timer function to take a raw ticks
value as the deadline instead of a monotonic time. This will make it
easier to reuse this function when scheduling boottime timers.

This required changing the pdev_timer_ops struct (and the RISC-V
bindings that use it) to also operate on raw ticks values.

Functionally, I expect this to be a no-op. I'm not changing the
computations performed, I'm just moving them up a layer in the stack.
The monotonic time -> raw ticks transformation that previously occurred
at the platform layer is now happening in the platform-agnostic code in
kernel/timer.cc.

Bug: 341785588
Change-Id: I9bc0fd8dfe811f009151ce75223220526ff92dde
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063733
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>