[build] Only include aml drivers & tests in arm64 builds
These drivers are for an arm64 SoC, and not useful in builds of
other architectures, but GN doesn't know that, and so includes
them, triggering the rebuilding of many artifacts in the arm64
toolchain.

This should save multiple minutes on a core.x64 build that
includes the buildbot tests.

NOTE: this doesn't move all the aml drivers, just a number of them, more
remain and need to be moved.

Change-Id: I7d64a718001c2a67cf7f93908668935d8c62579c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065706
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>