[kernel][pmm] Only check allocated
Corrects a bug in AllocPages where pages already in the list provided
had the pmm checker run on them, in addition to the pages actually
being allocated by the method. As there is no guarantee on the contents
of these pages, this could lead to the pmm checker spuriously failing.

Change-Id: I6bfb0054c377a1df782e56ee36fe538de1e13f25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1067132
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>