[fdf] Merge BootupWatcher to development Manager
Merge the BootupWatcher into the Manager protocol to
avoid the need of exposing the BootupWatcher service
everywhere. The BootupWatcher usage is in the same
spots as the Manager protocol so it's better to just
combine them into one.

Change-Id: I1e2a84066382603fc632b6aeb75128af15529b32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065957
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>