[starnix][selinux] Implement allow/deny-unknown class/permissions
Fix the policy index and permission checking implementations to
respect the policy-defined behaviour for unknown classes and
permissions.

This allows unknown permissions to be implicitly allowed or
denied, in addition to the previously-implemented behaviour of
rejecting the policy if it omits any kernel-defined permissions.

Redundant entries from micro-policies are removed.

Bug: 331197737
Change-Id: Ibf2b05921fd3b416d4621d9c54b722d213b7d5bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1056436
Commit-Queue: Wez <wez@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>