[fidl][abi_compat] Allow protocol openness mismatch
It's okay to change protocols between open, ajar and closed. The
problems are when you add and remove things when one of the peers thinks
that's not allowed, which is handled elsewhere.

This adds a small test for this case which required a bunch of changes
to support testing protocol-level skew.

Bug: 335446415
Change-Id: Ie3f3b2dc4fd40c978320b0273f14ab8a5eae62cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065801
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>