[bazel_sdk] Move SDK export rules to //build/bazel/bazel_sdk
Disable .pyc generation for the module at runtime following
a sys.path.insert() operation, this avoid polluting the Bazel
execroot with un-necessary files.

Also move the sdk_export script and targets to
//build/bazel/bazel_sdk since they are technically an
implementation detail of the Fuchsia platform build, and
do not need to be distributed with the Fuchsia Bazel SDK
out-of-tree.

Change-Id: I6221c181ca6fc034b2041bd26ffdbcb89bf4c283
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1067232
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: David Turner <digit@google.com>