[test][core] Clean test phys vmo
Performs a cache clean of the test physical VMO before returning it to
the calling test. This ensures it's in a known state and no previous
test has left data in the cache that might get written out as a cache
bomb if the current test is using the VMO uncached.

Change-Id: Ie730c833e60df143086809b0353c87dcf37e4341
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069592
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>