[lowpan] update discovery proxy workaround
The Fuchsia mdns do not support subscribe service via a specific network
interface. To make lowpan work with that, there was a workaround to keep
track of a list of addresses that both presented in the OT and
lowpan0. One case missed is that netstack added the address which
already presented in the OT, and that caused missing addresses in the
list and resulting discovery proxy may unable to work with locally
published services.

Bug: b/341965925
Change-Id: I12ccdf20b920c760183a11284b2da7b36227bd60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069779
Fuchsia-Auto-Submit: Jiaming (Charlie) Wang <jiamingw@google.com>
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>