[ld] Provide ubsan-custom runtime for test modules
The test modules can be built in the standalone ubsan variant,
while fancier sanitizer variants like asan are already excluded.
It just needs a custom runtime.  On POSIX, this can be complete.
On Fuchsia, it has no way to print but it formats a message on
the stack and crashes in a way where the pointer to the message
buffer can be found in a known register with a debugger.

Bug: 348678677
Change-Id: I355bcb42ff52b8ab38d2612956e3c342f49ce2e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1066593
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Leonard Chan <leonardchan@google.com>