[build] Use new "custom-runtime" toolchain tag
This marks a variant like ubsan, that doesn't have
"instrumentation-runtime" because everything in a special
environment should wire up a custom runtime (in the case of
ubsan, via //src/lib/ubsan-custom) if it's feasible.  Things
where it's not feasible to usefully wire up any runtime can
exclude the variant on the basis of "custom-runtime" as well as
excluding "instrumentation-runtime".

This is now used for the things that are built in the user.basic
environment but cannot wire up ubsan-custom.

Bug: 348678677
Change-Id: I959d3ad2063bad95d788ee87c3445063dc3d68f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065936
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>