[netstack3] Remove datagram sharing update race
The form of `update_sharing` in the datagram API allowed for potentially
racy behavior in the UDP implementation. This is not really an
observable race because bindings linearizes all the operations on the
socket, but this new form makes ite clearer that we shouldn't have races
here.

Fixed: 345740748
Change-Id: I7eb8634314b70a534eb6c168e729435fae6386bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069783
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>