[build] Support FUCHSIA_BAZEL_DISK_CACHE environment variable.
Allow the user to set an environment variable to point to an
absolute directory path that will be used to store build
artifacts out of the Ninja build directory.

This is useful when full Bazel remoting is not enabled,
(i.e. when `bazel_enable_rbe = true` it not set in args.gn),
when performing builds on different checkouts of the same
machine, or when doing repeated clean builds, since the
Fuchsia platform build does not use the standard output_base
locations.

While the Bazel documentation states that using --disk_cache
is compatible with remote caching, the RBE documentation
states that this flag 'does not mix' with RBE and thus
should be removed from the `.bazelrc` file, so avoid using
the flag when `--config=remote` is used.

Change-Id: Ie15e9b1b1c11c1f700bb2e67f919e0618acd03a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1070574
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>