[Honeydew] Replace subprocess.* with host_shell.*
* subprocess.check_output() has been replaced with host_shell.run()
* subprocess.check_call() has been repalced with host_shell.run()
* subprocess.Popen() has been repalced with host_shell.popen()

We have many places in Honeydew that performs subprocess oeprations and
handles the exceptions. All of this code is being repeated in multiple
places. So created host_shell module with all subprocess related
operations and replaced all of subprocess.* calls in Honeydew with
host_shell* calls.

As part of this, also removed 'exceptions_to_skip' argument used in all
ffx.run() and fastboot.run() as it is not being used anywhere (except
for one usecase which can be handled differently) and is making the code
less readable.

BUG: b/347778716, b/332538317

Change-Id: If7b2aea4d8879736ba551bcab6bb6ff77f78f342
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065893
Commit-Queue: Pavan Kumar Juturu <jpavankumar@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>