[test][core] Correct cpu stats test
This corrects a test that is attempting to validate that the number of
records of the ZX_INFO_CPU_STATS can be queried without needing to
provide a buffer. However, the test was doing this by providing the
nullptr for the buffer with a non-zero length. This happened to work
due to the passed in length being less than the size of a single record
and so the kernel not actually attempting to perform any copies to
realize that the buffer was invalid.

Ideally the kernel would validate every passed in buffer range, but
this cannot be efficiently done, and the test was therefore relying on
what is essentially undefined behavior. This commit changes to a
correct buffer size of 0, which will allow the call to reliably work.

Change-Id: Ibc4f07e88a97b1de12e0d0878735d00e232780e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069572
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>