[kernel][phys] Introduce memalloc::Pool::RestrictTotalRam()
This method will be the backbone of the (re)implementation of the
kernel.memory-limit-mb boot option. It removes all accounting for RAM
(free or allocated) past the provided capacity. It is a destructive
routine; removed memory can not be recovered or reallocated.

It is always the tail of RAM that is removed. If allocations are
removed, then that memory is simply forgotten about by the pool. This
applies to bookkeeping to an extent as well: the tracking of bookkeeping
allocation may be removed, but the pool might still have access to
bookkeeping in forgotten ranges, possibly continuing to draw from it
rather than allocate new tracked bookkeeping ranges. This is okay, as it
is assumed that memory that the pool does not track will not be used by
any program.

Bug: 347766366
Change-Id: Idf7d3c34c0fe63c7c41f5579807dfbb43510c3ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1070845
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>