[netstack3] Fix deferred TCP socket destruction
The logic for marking a socket for DOA was inverted: if a socket is *not
yet* marked for destruction it must be inserted into the sockets set.

Added details to loom test to cover this, which required exposing
deferred destruction through `FakeBindingsCtx`.

Bug: 349404597
Change-Id: I812ad36f26add457b4ea4883ac6d0eb9dd6fb505
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1071839
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>