[kernel] Give a simpler handling of kernel.memory-limit-mb
We do not need the complexities of libmemory_limit to handle this boot
option. With a normalized accounting of RAM, we can more simply
determine a RAM cut-off address and then truncate all mention of RAM
passed that point.

Removing the use of libmemory_limit, which parameterizes repeated calls
to pmm_add_arena() is a first step toward refactoring how the arenas are
initialized (which should happen all at once, and not an arena at a
time).

The code in this CL is a switchback, tweaking the legacy zbi_mem_range_t
ranges - as described by the bootloader - that are handed off to the
kernel. When we are ready to hand off the normalized ranges, we'll be
able to remove this special handling entirely, relying on the truncation
of the normalized ranges that now occurs at the very beginning of phys
memory intialization.

Bug: 347766366
Change-Id: I2460fdd6e941a596dd7a6daf4f3b2aacdf026c37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1067512
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>