[zircon] Clarify state of FIDL library `zx`
Attempt to make it clearer that the library is not experimental and is
in a stable but interim state.

In addition, clarify the difference between the two `zx` libraries,
including adding a README for the `zx/` subdirectory.

Also update and fix links in the existing README.

Change-Id: I35c2db6418530645effaa78a820c87af60c10985
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051240
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>