[session] make cobalt unit tests more robust
Configure the unit tests to panic if the executor stalls.

This makes the tests fail-fast in some conditions where they would
otherwise wait forever.

For example: suppose we introduce a bug in `log_session_launch_time()`,
which causes that function to believe that the `end_time` in the
`test_log_session_launch_time_swap_start_end_time` test is greater
than the `start_time` in that test.

With the previous code, the test would hang forever. This is because
`log_session_launch_time()` would be waiting for a reply to its
call to `log_integer()`, but the test does not have any affordance
to return a reply. (And the test _shouldn't_ have such an
affordance, because the test is not expecting such a call to be
made.)

Test: fx test session_manager_tests
Change-Id: I0fb6d32350fe699a8b3234620468f89923d572b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069534
Reviewed-by: Caroline Liu <carolineliu@google.com>
Commit-Queue: Mukesh Agrawal <quiche@google.com>
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>