[starnix][kernel] Clarify get_waitable_child result
Previously, this function returned a `Result<Optional<...`, which made
it unclear what the difference was between an error result and an
`Ok(None)` result. This CL changes the function to return a custom enum
that makes the `ShouldWait` and the `NoneFound` cases clear.

This clarity is especially useful for the one caller, which previously
intercepted the `Err(...)` result and sometimes did not generate an
error (i.e., in the case where it had a tracee to wait for).

Change-Id: I9100e5fb4fbb456b2f852c8c54fdf0122e61db29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1072852
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>