[iob][tests] Fix flaky test
Fixes a test that attempts to query all the mappings in a process, but
allocates between determining the number of mappings, and performing the
query, potentially increasing the number of mappings due to heap growth.
If the heap was placed, due to aslr, below where the IOB mapping was
placed, and the IOB mapping is virtually the last mapping in the
process, this will cause the IOB mapping to not be in the list.

Change-Id: Ifcffbbcaea3e3d6c4afbb12a2a7123ae13e52f6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073112
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>