[power,test] Share code between suspend timing and power metrics
Share code between the two metrics collectors that capture information
about suspend and resume. This centralizes the place where we hard-code
things like the names and categories of trace events on which we rely.

The code that calculates power usage during suspend remains sequestered
inside the PowerMetricsCollector so as to avoid having knowledge of
the current power/trace synchronization strategy leak outside that
class.

Bug: 349601252
Test: fx test power_metrics_test
Change-Id: If07302c384254c50da8d03f89fa2b6f2fa8659d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1072773
Commit-Queue: Chris Masone <cmasone@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>