[components][sandbox] Replace OneShotHandle with zx.Handle
In the Capability union, handles are represented by a OneShotHandle type
that is defined by the sandbox fidl. Dictionary requires that all items
can be infallibly cloned, but this is not always true for handles. The
OneShotHandle type was invented to get around this - as the name
implies, it employs the "oneshot" pattern to make handle capabilities
always copyable, exposing a Take method that either extracts the
interior handle or returns null.

This change replaces OneShotHandle with a raw zx.Handle. This has a
number of advantages, in particular it removes a layer of abstraction
and is a better fit for handles that are duplicable. To accommodate
this, we need to update the interface of some Dictionary methods to
allow them to return a NOT_CLONEABLE error, or (in the case of
Enumerate) null for values that could not be cloned. Fortunately, the
overall impact on the clients of Dictionary as a result of this change
is small and manageable.

Fixed: 339492452

Change-Id: I30df0a74592e427d0c67584c598158d59054dabe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052993
Reviewed-by: David Gilhooley <dgilhooley@google.com>
API-Review: Hunter Freyer <hjfreyer@google.com>
API-Review: Gary Bressler <geb@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Gary Bressler <geb@google.com>