[starnix][vfs] Fix crash when prerun fails
When the prerun function fails, we were crashing when unwinding the
thread in some situations because the task_complete closure was holding
some releasables. When we dropped those objects, they would register
themselves in the RELEASABLES thread local but never actually get
released.

Now, we finalize the DelayedReleaser to catch those objects when they
try to register themselves instead of catching them when we tear down
the thread.

Also, we drop the task_complete closure in the error codepath to drop
the releasables while the thread local is still able to accept them.

Change-Id: I00acef7dda5fd4030a6f62cc4ec8344f1fc23629
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073480
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>