[routing] move with_policy_check from component manager
The `with_policy_check` function will need to be invoked by
`cm_fidl_analyzer`, so this commit moves it to the routing library in a
new trait.

This function requires the ability to get a moniker from a
`sandbox::WeakComponentToken`, which was not possible to do before in a
generic way due to the type erasure perfomed by `WeakComponentToken`. To
address this, a new field `moniker` has been added to the struct.

Change-Id: I7db5d1ed6883e1a84d0ab8ce2f2b68eb35e6d152
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1059752
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Gary Bressler <geb@google.com>