[tefmocheck] Create new tefmo for seriallistener timeout failures.
The Boot-test-successful check is to capture cases where the boot test
passed but we failed to detect it and thus failed the test. However, if
a boot test legitimately times out because it doesn't see the success
string, it will print out that it fails to find the success string
Boot-test-successful in the swarming output. The original check tracks
both of these types of failures together, but we should separate the two
since they are two different types of failures.

NO_IFTTT=just added the check to keep the strings in sync
Bug: 42067738
Change-Id: I094567df6997b31c451d27e3a2a412682a5e345e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064833
Reviewed-by: Danielle Kay <danikay@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>