[starnix][tests] Run reboot test in debian container
This CL migrates the reboot test to run in the debian container. To make
this work, we need to write our own client for the reboot syscall
instead of relying on one from the container.

I tried using the `/sbin/reboot` binary in the Alpine container, but
that binary just sends a signal to `init` rather than calling the reboot
syscall.

Change-Id: Ia170ea66967d6d985b78aea715e6cd6ed6fed846
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073372
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Adam Barth <abarth@google.com>