[kernel][vm] virtual_alloc better contiguous check
Exposes the internal check VirtualAlloc was doing to decide whether to
actually perform contiguous large page mappings and has the unittests
also check that when determining if a contiguous allocation can be
expected.

Additionally the CanExpectContiguous method is extended to place the
PMM in a state such that the, currently disabled, tests would reliably
fail if run unmodified, instead of 'accidentally' succeeding.

Bug: 349402040
Change-Id: I2b346a0faaebe57311d647aed6db2590eea35e70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073092
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>