[aml-uart] Request interrupt based on driver config
The driver registers the interrupt it requests as wake source without
considering the driver config, with this change, the driver will only
register the interrupt as wake source when suspend is enabled in driver
config.

Test: Verified bluetooth functionality on Smart Display and OOBE worked
well, verified suspend and resume on VIM3 and it worked as expected.

Change-Id: I496b4fcfa87e5be593ac78c5be6cd13302061a11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073652
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>