[debugger] Add TestFailureStackMatcher with tests
This cleans up the code that detects when we are running embedded in fx
test and allows ConsoleContext to have enough external context to tell
when we can jump straight into the user's code from a test failure.

The tests use some real stack frames from failed gTest and Rust tests
and show that we can identify the top frame of boilerplate. This is done
by mimicking the namespace hierarchy with fake symbols to create the
same symbol names for the test.

Change-Id: I49a16b8122f97f2cc61749d0e6de0ba5e0d28c84
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074257
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>