[audio] Don't set a default dispatcher
In the future, logging will use the default dispatcher if one
is available. In drivers, where a dispatcher can be terminated,
this doesn't make sense. This patch disables the default dispatcher,
preventing logging from attempting to use it after it's been shutdown
(essentially a use-after-free), resulting in a crash. Not
setting the default dispatcher causes the logger to maintain
its existing behavior of spawning a thread, resulting in no crash.

Bug: b/349909461
Change-Id: Ifd4d37d0e7583d5c47779280feaf57557c7e4a3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074513
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>