[netstack3] Reduce lock acquisition in TCP connect
This CL achieves 2 goals

* Reduce demux lock in `connect` to a smaller critical section, which
  doesn't require calling into the IP layer while holding the demux
  lock.
* Reduce the number of match arms to call connect_inner, since the
  natural case is 3 arms: v4->v4, v6->v4, v6->v6. The CL lifts the
  problem of when and how to acquire the opposite demux lock to a new
  helper accessor struct and trait.

To achieve that, we introduce `DualStackTuple` which allows us to more
freely use the pattern <SockI, WireI> and convince the compiler that a
dual stack context on SockI should be able to fulfill any WireI that
we're given.

Bind is next, with the goal to remove the context functions allowing
simultaneous access to demux and IP layer.

Bug: 327489613, 327488712
Change-Id: Ia3e48935506c7e51edf00155cf8ec4563e4c3505
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074274
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>