[rbe] Rename to 'legacy_default'
This makes it a little clearer that 'legacy_default' (resembling the old
default from the Goma-only days) is not a choice most users want,
even if it is what they are used to.
We want to encourage users to pick a baseline mode that
best matches their development environment.

Bug: 344673187
Change-Id: Ie76a6e4427ecef9ea018226996e94846ba3293c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074465
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: David Fang <fangism@google.com>