[versioning] Print Git instructions in freeze/bump scripts
This makes it clear which files are changed, especially if we need to
run these steps manually. Running these exact commands also avoids
committing files that were not expected to be changed in this process.

Move the instructions for updating availability_levels.inc to before the
bump instructions.

Bug: 349204008
Change-Id: Ie7775a67181d220b98f26a181b31438db6dd7f46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1071045
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>