[perfcompare] Disable error from "--expected_sample_size" for easier rollout
Turn the error into a warning.  This means we can first change the
Infra recipe to enable passing "--expected_sample_size" and then
enable the error with a change on the fuchsia.git side.

That makes for easier testing and rollout.  The CQ will run internal
builders for fuchsia.git CLs but it only runs public builders for
recipes.git CLs, which makes it harder to test recipes.git changes.

Bug: 343997506
Change-Id: I0cab509c167e10acd3272365ae9b3840340b44b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074512
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>