[idk] Rename override_target_api_level
Rename GN arg `override_target_api_level` to
`current_build_target_api_level`, which more accurately describes its
purpose since https://fxrev.dev/1028476. Specifically, it is the API
level that the current GN build is targeting. It is overridden for
sub-builds, but it is not expected to be overridden in other contexts.

Bug: 334936990
Change-Id: Ifc017bf91c1cda3dc0f1d42fd26ae8bbbdc85b62
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073485
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>