[kernel][phys] s/"extended"/"allocated" in libmemalloc
We initially dubbed the allocated range types as "extended" owing to
their existence as those that live in the extension of the
zbi_mem_type_t value space. This is less clear of a name though; better
to just call them "allocated", especially as we begin to start using
this concept in the kernel proper.

Bug: 347766366
Change-Id: I398630edde0ccf9c46db8f419011e47fcb580baa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1071902
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>