[versioning] Replace bump-related asserts with messages
https://fxrev.dev/1062737 added asserts to enforce the conditions that
must be true in order for the simplified implementation to be correct.
Although these are true during the build, they are not true if the GN
rules are run after the build; specifically, after the bump build adds
the new level to version_history.json.

fint-based builds, including platform-version-roller, do this when they
call gnCheckGenerated().

Convert the assert to a warning message that indicates something is
wrong but can be ignored on the bot in this case. The need for these
will be eliminated in upcoming CLs related to the NEXT API level.

Bug: 305961460
Test: `fx set ... --args=bump_api_level=true && fx build && fx gn check out/default`
Change-Id: I7747fcd650a7c7bb6c409c2eb23c1947306e0a8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1075392
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>